Delay runit restart on config file change
On first install, the consul service may not exist when we create the
`node[:consul][:config_dir]}/default.json`. This leads to an error as
the service tries to restart itself (thanks to `subscribes`) before the
`/etc/init.d/consul` is created.

This parallels the init.d change d1f21164d03f2cf74b995e61219565821189b5f3 by jubianchi
authored on Jul 25 but for runit