Fixes #262 for Windows
It appears the nssm resource also isn't idempotent. I also added
another attribute and made it the default value for the install_path
property for consul_service. My reasoning is that one should be able
to change this crucial property without having to redeclare the whole
resource. #262 will be fixed for linux once
https://github.com/johnbellone/libartifact-cookbook/pull/2 is merged