Removed problematic reload line
`new_resource.notifies(:reload, new_resource, :delayed)` causes the
`consul_service` resource to reload every time Chef ran. This is
an issue with the new sysvinit service provider when the service is
in a stopped state when Chef runs. It would start and recieve a reload
HUP too quickly & subsequently crash the service.