Fix #288 - No longer restart Consul automatically.
This seems to have come up a few times in the recent past. The idea here
is that we do not want to update Consul because the Chef Client may be
running in parallel on all of the nodes and result in a Consul cluster
downtime. This separately also sends a SIGHUP (reload) when a
configuration change is made instead of restarting the agent (for the
same reasons).

Because of this change it means that there will no longer be seamless
upgrades. It will require an operator to restart the Consul process on
all of the machines. Or they're going to need to wait until the instance
restarts since it'll be picked up at the next boot.