Remove default value for `ui_dir`
Needs to be unset for the `ui` flag to work as expected.  The `ui_dir`
flag overrides the embedded web UI.  If set, and the web UI is not
available at the specified path, the server will return 404s.

A default value in the resource means that even if the node attribute is
explicitly rm'd, the config value will still be populated.

Fixes #339, refs #315