Remove the notification for consul service restart
We don't need to notify consul service by changes of `poise_service_user` resource:

- It is useless for initial converge because the service will start automatically after and
it happens _after_ user creation.

- We should not restart the service automatically even if the name of service user has been
changed for an existing installation. In this case the restart is harmful, because consul data
(/var/lib/consul/*) is still owned by an old user, so it should be handled by a special order:
1) stop service, 2) change data files ownership, 3) start service.

So, it's out of this cookbook's scope.