Restore the ability to set a custom open file limit
This seems to have existed, at least for Upstart, in [older versions](https://github.com/sous-chefs/consul/blob/v0.11.1/attributes/default.rb#L123) of the cookbook, but disappeared with the conversion to poise-service.

The Kitchen suites have a pretty basic default config that I was hesitant to mess with, hence the new test suite, which could also be used to validate other config overrides in the future. If the additional build time is going to be a problem, I can rip those tests back out or put them in the default suite.

Signed-off-by: Jonathan Hartman <j@hartman.io>